+++ wp-content/plugins/shola-core/shola-core.php
@@ -52,6 +52,7 @@ spl_autoload_register( 'shcore_autoload' );
 \SholaCore\Meta_Fields::init();
 \SholaCore\Contact_Settings::init();
 \SholaCore\Social_Links_Settings::init();
+\SholaCore\View_Counter::init();
 \SholaCore\Roles::init();
 \SholaCore\SEO::init();
 \SholaCore\Security::init();